broken filo spacing - start of big update
Began adding functionality for filo spacing. This builds on the existing functionality for spacing of cortical puncta (using fluorescence intensity on the cell perimeter band).

In this version I tried to add filos to the perimeter using white pixels - but there is a bug in the index variable so the crossing solutions are wrong (should be "i" instead of "index"). Futher more it isnt necessary to use the find maxima plugin since we already know the crossing coordinates.