Analysis version 8
complete filo spacing functionality - changed analysis version from 7 to 8

in this version the ROIs for filo-cell contacts are estimated using the tip-cell vector with an optional step for the user to adjust these ROIs. The ROIs are now saved in a file called filo-spacing-RoiSet8. Previously these ROIs were displayed but not edited or saved.

These ROIs are used to calculate the intersection point with the cell perimeter resulting in several new files:
* filo-spacing-x-x.tif has the band image converted to 16-bit with a white pixel marking each filo crossing point.
* filo-spacing-x-x.txt has the nearest neighbor spacing distance for each filopod (compare with the previous file)
* filo-spacing-results-8.csv has summary statistics for the filo spacing
* if debug mode is enabled there will be a couple of additional files - the ROIs from the crossing solutions and an unmasked image with ROIs drawn on it. The results table will also be populated with raw data from the crossing solution when in debug mode.

Currently not working: the cell perimeter analysis