fixed cortical spacing analysis
Several fixes to the banded spacing analysis. The boxheight parameter (now standardized since being alternately called boxheight or boxwidth) was larger when the band spacing analysis was first tested. The find maxima function was not detecting any maxima with the current thin setting of the boxheight. This was corrected by allowing the plugin to find edge maxima. Consider using a larger boxheight and turning exclude back on.