Filopod length support + bug fix for degenerate solutions
Calculate filopod length directly from the crossing point. There was a bug in the previous commit where degenerate solutions (180 degrees opposite on the cell perimeter) would be reported if they were encoutered first. Now the solutions are check to be sure they fall in the fp bounding rectangle as well as the cell border rectangle. Another test for degeneracy is to prefer the solution with shorter length but I am not sure this would work in all cases and it is redundant here, so I have left it commented out.